pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "quack-kernels"
dynamic = ["version"]
requires-python = ">=3.9"
dependencies = [
    "nvidia-cutlass-dsl==4.1.0.dev0",
    "torch",
]

[project.optional-dependencies]
dev = [
    "pre-commit",
    "ruff",
]

[tool.setuptools.packages.find]
exclude = ["tests", "benchmarks"]

[tool.setuptools.dynamic]
version = {attr = "quack.__version__"}

[tool.ruff]
line-length = 100

[tool.ruff.lint]
ignore = [
    "E731",  # do not assign a lambda expression, use a def
    "F841",  # local variable is assigned to but never used
]